# HuaHuoSwagger/swagger.json
{"openapi":"3.0.1","info":{"title":"Backend of HuaHuo APIs.","version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"security":[{"JWT_AUTH":[]}],"paths":{"/binaryfiles":{"get":{"tags":["binary-file-controller"],"operationId":"listBinaryFiles","parameters":[{"name":"pageNumber","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"isElement","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListBinaryFileResult"}}}}}}},"/binaryfiles/{fileId}/coverPage":{"post":{"tags":["binary-file-controller"],"operationId":"uploadCoverPage","parameters":[{"name":"fileName","in":"query","required":true,"schema":{"type":"string"}},{"name":"fileId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"isElement","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"coverPageFile":{"type":"string","format":"binary"}}}}}},"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryFileCallStatus"}}}}}}},"/login":{"post":{"tags":["login-controller"],"operationId":"login","parameters":[{"name":"userName","in":"query","required":false,"schema":{"type":"string"}},{"name":"password","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginStatus"}}}}},"security":[]}},"/users/exist":{"get":{"tags":["user-controller"],"operationId":"exist","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserExistResponse"}}}}}}},"/binaryfiles/{fileId}/description":{"put":{"tags":["binary-file-controller"],"operationId":"updateBinaryFileDescription","parameters":[{"name":"fileId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"text/plain":{"schema":{"type":"string"}}}},"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryFileCallStatus"}}}}}}},"/binaryfiles/{fileId}":{"get":{"tags":["binary-file-controller"],"operationId":"downloadBinaryFile","parameters":[{"name":"fileId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"application/octet-stream":{"schema":{"$ref":"#/components/schemas/Resource"}}}}}},"delete":{"tags":["binary-file-controller"],"operationId":"deleteBinaryFile","parameters":[{"name":"fileId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/users":{"get":{"tags":["user-controller"],"operationId":"all","responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDB"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","parameters":[{"name":"isAnonymous","in":"header","required":false,"schema":{"type":"boolean"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDB"}}}},"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDB"}}}}},"security":[]}},"/elements/{elementId}":{"get":{"tags":["element-controller"],"operationId":"getElement","parameters":[{"name":"elementId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ElementDB"}}}}}}},"/binaryfiles/upload":{"post":{"tags":["binary-file-controller"],"operationId":"uploadFile","parameters":[{"name":"fileName","in":"query","required":true,"schema":{"type":"string"}},{"name":"forceOverride","in":"query","required":false,"schema":{"type":"boolean","default":true}},{"name":"isElement","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryFileCallStatus"}}}}}}},"/binaryfiles/existFile":{"get":{"tags":["binary-file-controller"],"operationId":"existFile","parameters":[{"name":"fileName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryFileExistResponse"}}}}}}},"/tokenValid":{"get":{"tags":["login-controller"],"operationId":"isTokenValid","parameters":[{"name":"userName","in":"query","required":true,"schema":{"type":"string"}},{"name":"jwtToken","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenValidResponse"}}}}}}},"/binaryfiles/coverPageNotFound":{"get":{"tags":["binary-file-controller"],"operationId":"coverPageNotFound","responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Resource"}}}}}}},"/users/{id}":{"get":{"tags":["user-controller"],"operationId":"one","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDB"}}}}}}},"/elements":{"post":{"tags":["element-controller"],"operationId":"createElement","parameters":[{"name":"elementId","in":"query","required":true,"schema":{"type":"string"}},{"name":"fileId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"editable","in":"query","required":false,"schema":{"type":"boolean","default":true}},{"name":"sharable","in":"query","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"404":{"description":"default response","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"default response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ElementDB"}}}}}}}},"components":{"schemas":{"BinaryFileDB":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"version":{"type":"string"},"createdBy":{"type":"string"},"description":{"type":"string"},"checksum":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"modifiedTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]},"fileType":{"type":"string","enum":["PROJECT","ELEMENT"]}}},"ListBinaryFileResult":{"type":"object","properties":{"totalCount":{"type":"integer","format":"int32"},"binaryFiles":{"type":"array","items":{"$ref":"#/components/schemas/BinaryFileDB"}}}},"BinaryFileCallStatus":{"type":"object","properties":{"fileId":{"type":"integer","format":"int64"},"succeeded":{"type":"boolean"},"msg":{"type":"string"}}},"LoginStatus":{"type":"object","properties":{"userName":{"type":"string"},"failReason":{"type":"string"},"jwtToken":{"type":"string"}}},"UserExistResponse":{"type":"object","properties":{"username":{"type":"string"},"exist":{"type":"boolean"}}},"UserDB":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"password":{"type":"string"},"nickname":{"type":"string"},"role":{"type":"string","enum":["ADMIN","CREATOR","READER","ANONYMOUS"]},"status":{"type":"string","enum":["ACTIVE","SUSPENDED","DELETED"]},"createTime":{"type":"string","format":"date-time"},"lastLoginTime":{"type":"string","format":"date-time"},"modifiedTime":{"type":"string","format":"date-time"}}},"ElementDB":{"type":"object","properties":{"elementId":{"type":"string"},"editable":{"type":"boolean"},"shareable":{"type":"boolean"},"binaryFileDB":{"$ref":"#/components/schemas/BinaryFileDB"}}},"BinaryFileExistResponse":{"type":"object","properties":{"fileName":{"type":"string"},"exist":{"type":"boolean"}}},"Resource":{"type":"object","properties":{"open":{"type":"boolean"},"file":{"type":"string","format":"binary"},"url":{"type":"string","format":"url"},"readable":{"type":"boolean"},"uri":{"type":"string","format":"uri"},"description":{"type":"string"},"filename":{"type":"string"},"inputStream":{"type":"object"}}},"TokenValidResponse":{"type":"object","properties":{"userName":{"type":"string"},"isValid":{"type":"boolean"}}}},"securitySchemes":{"JWT_AUTH":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}